# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Dscommerce (E-commerce) API","description":"API REST for e-commerce. OAuth2 authentication with JWT.","license":{"name":"MIT","url":"https://github.com/mateusribeirocampos/project-spring-boot-dscommerce"},"version":"v1.0.3"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Order","description":"Controller for order"},{"name":"Password Recovery","description":"Controller for password recovery"},{"name":"Category","description":"Controller for categories"},{"name":"Product","description":"Controller for product"},{"name":"Users","description":"Controller for users"}],"paths":{"/oauth2/token":{"post":{"tags":["Authentication"],"summary":"Get / Refresh Access Token","description":"Supports two grant types: `password` (login) and `refresh_token` (token rotation).","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"oneOf":[{"required":["client_id","client_secret","grant_type","password","username"],"type":"object","properties":{"grant_type":{"type":"string","enum":["password"]},"client_id":{"type":"string","example":"myclientid"},"client_secret":{"type":"string","example":"myclientsecret"},"username":{"type":"string","example":"maria@gmail.com"},"password":{"type":"string","example":"123456"}},"description":"Password grant (login)"},{"required":["client_id","client_secret","grant_type","refresh_token"],"type":"object","properties":{"grant_type":{"type":"string","enum":["refresh_token"]},"client_id":{"type":"string","example":"myclientid"},"client_secret":{"type":"string","example":"myclientsecret"},"refresh_token":{"type":"string","example":"<refresh_token_value>"}},"description":"Refresh token grant (rotation)"}]}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"},"token_type":{"type":"string","example":"Bearer"},"expires_in":{"type":"integer"}}}}}},"401":{"description":"Invalid credentials"}},"security":[]}},"/users/me":{"get":{"tags":["Users"],"summary":"Profile","description":"Controller for visualize user profile","operationId":"getMe","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"put":{"tags":["Users"],"summary":"Update profile","description":"Controller for update user profile","operationId":"updateMe","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/products/{id}":{"get":{"tags":["Product"],"summary":"Find product by id","description":"Controller to find product by id","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"put":{"tags":["Product"],"summary":"Update product by id","description":"Controller to update product by id","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"delete":{"tags":["Product"],"summary":"Delete product by id","description":"Controller to delete product by id","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"201":{"description":"Created"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/orders/{id}":{"get":{"tags":["Order"],"summary":"Find order by id","description":"Controller to find order by id","operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"put":{"tags":["Order"],"summary":"Update order by id","description":"Controller to update order","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"delete":{"tags":["Order"],"summary":"Delete order by id","description":"Controller to delete order by id","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/categories/{id}":{"get":{"tags":["Category"],"summary":"Find category by id","description":"Controller to find category by id","operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"put":{"tags":["Category"],"summary":"Update category by id","description":"Controller to update categories","operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"delete":{"tags":["Category"],"summary":"Delete category by id","description":"Controller to delete category by id","operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/users/register":{"post":{"tags":["Users"],"summary":"Register","description":"Controller for register users","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInsertDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/products":{"get":{"tags":["Product"],"summary":"Find all products","description":"Controller for visualize all the products","operationId":"findAll","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagedModel"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"post":{"tags":["Product"],"summary":"Insert product","description":"Controller to insert product","operationId":"insert","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/orders":{"get":{"tags":["Order"],"summary":"Find all orders","description":"Controller to find all the orders","operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDTO"}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"post":{"tags":["Order"],"summary":"Insert order","description":"Controller to insert order","operationId":"insert_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/categories":{"get":{"tags":["Category"],"summary":"Find all categories","description":"Controller to find all the categories","operationId":"findAll_2","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDTO"}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}},"post":{"tags":["Category"],"summary":"Insert category","description":"Controller to insert category","operationId":"insert_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDTO"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}},"/auth/reset-password":{"post":{"tags":["Password Recovery"],"summary":"Reset Password","description":"Controller for reset password","operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewPasswordDTO"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/auth/forgot-password":{"post":{"tags":["Password Recovery"],"summary":"Forgot Password","description":"Controller for password recovery","operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDTO"}}},"required":true},"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StandardError"}}}}}}}},"components":{"schemas":{"StandardError":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"}}},"UserUpdateDTO":{"required":["email","name"],"type":"object","properties":{"id":{"type":"integer","description":"Database generated User ID","format":"int64","readOnly":true},"name":{"maxLength":80,"minLength":3,"type":"string","description":"User name"},"email":{"type":"string","description":"User email"},"phone":{"type":"string","description":"User phone"},"birthDate":{"type":"string","description":"User birth date","format":"date"},"roles":{"type":"array","description":"User role","readOnly":true,"items":{"type":"string","description":"User role","readOnly":true}}}},"UserDTO":{"required":["email","name"],"type":"object","properties":{"id":{"type":"integer","description":"Database generated User ID","format":"int64","readOnly":true},"name":{"maxLength":80,"minLength":3,"type":"string","description":"User name"},"email":{"type":"string","description":"User email"},"phone":{"type":"string","description":"User phone"},"birthDate":{"type":"string","description":"User birth date","format":"date"},"roles":{"type":"array","description":"User role","readOnly":true,"items":{"type":"string","description":"User role","readOnly":true}}}},"CategoryDTO":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":50,"minLength":3,"type":"string"}}},"ProductDTO":{"required":["description","imgUrl","name","price"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":80,"minLength":3,"type":"string"},"description":{"maxLength":2147483647,"minLength":50,"type":"string"},"price":{"type":"number","format":"double"},"imgUrl":{"pattern":"^https://.*","type":"string"},"categories":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDTO"}}}},"ClientDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"OrderDTO":{"required":["items"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"moment":{"type":"string","format":"date-time"},"orderStatus":{"type":"string","enum":["WAITING_PAYMENT","PAID","SHIPPED","DELIVERED","CANCELED"]},"client":{"$ref":"#/components/schemas/ClientDTO"},"payment":{"$ref":"#/components/schemas/PaymentDTO"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDTO"}},"total":{"type":"number","format":"double"}}},"OrderItemDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number","format":"double"},"quantity":{"type":"integer","format":"int32"},"imgUrl":{"type":"string"},"subTotal":{"type":"number","format":"double"}}},"PaymentDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"moment":{"type":"string","format":"date-time"}}},"UserInsertDTO":{"required":["email","name","password"],"type":"object","properties":{"id":{"type":"integer","description":"Database generated User ID","format":"int64","readOnly":true},"name":{"maxLength":80,"minLength":3,"type":"string","description":"User name"},"email":{"type":"string","description":"User email"},"phone":{"type":"string","description":"User phone"},"birthDate":{"type":"string","description":"User birth date","format":"date"},"roles":{"type":"array","description":"User role","readOnly":true,"items":{"type":"string","description":"User role","readOnly":true}},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"NewPasswordDTO":{"required":["password","token"],"type":"object","properties":{"token":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"ForgotPasswordDTO":{"required":["email"],"type":"object","properties":{"email":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModel":{"type":"object","properties":{"content":{"type":"array","items":{"type":"object"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}